+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2021-08-19
+
+### Added
+- Reset method for scopes
+- Ring example
+
+### Fixed
+- Debug output
+
+## [0.1.0] - 2021-08-19
 ### Added
 - Re-exports of allocator-api types and traits when "nightly" feature is enabled.
 - Copies of allocator-api types and traits when "nightly" feature is disabled